+++ app/controllers/recording_studio_billing/admin_options_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RecordingStudioBilling
+  class AdminOptionsController < AdminPageController
+    def new
+      render_billing_admin_page(:billing_option_new)
+    end
+
+    def edit
+      render_billing_admin_page(:billing_option_edit)
+    end
+  end
+end