+++ app/controllers/recording_studio_billing/admin_prices_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RecordingStudioBilling
+  class AdminPricesController < AdminPageController
+    def new
+      render_billing_admin_page(:price_new)
+    end
+
+    def edit
+      render_billing_admin_page(:price_edit)
+    end
+  end
+end